requirements.txt
Flask==2.0.1
Flask-AppBuilder==3.3.2
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
gunicorn==20.1.0
python-dotenv==0.19.0
viberbot==1.0.11